requirements.txt
flask
pandas
playwright
python-telegram-bot==20.3